file restructuring
restructured the files to fit neatly into backend and frontend folders
