+++ backend/core/conversation-manager.js
