+++ backend/core/message-handler.js
