+++ backend/providers/code-assistant-provider.js
